Save to SVG, and to PDF and SVG
- Save to SVG

- Save to PDF and SVG (useful for our books). Specify the base file
  name (e.g. foo) and get foo.svg and foo.pdf

- Removed PdfCanvas and replaced it with FileCanvas, which provides save
  to PDF, SVG, and PDF and SVG.
