+++ shared/src/main/scala/doodle/backend/Formats.scala
@@ -6,5 +6,6 @@ object Formats {
   sealed trait Gif
   sealed trait Svg
   sealed trait Pdf
+  sealed trait PdfAndSvg
   sealed trait Screen
 }
